Remove +[RACScheduler deferredScheduler]
When used from a background thread without a scheduler of its own, the deferred scheduler becomes dangerous. It won't actually defer execution until _after_ the calling code -- because it dispatches to a different queue, the scheduled block may run concurrently to the caller, which would be wholly unexpected.

It's much better to explicitly use +mainThreadScheduler, +currentScheduler, or +backgroundScheduler instead.