Refactor +subscriptionScheduler to default to a background scheduler
Although the main scheduler is probably safer, this has a high likelihood of unintentionally pushing a lot of work onto the main queue (basically any time a subscription happens from a background GCD or operation queue), and we've never guaranteed that subscription happens on the main thread by default anyways. This will continue to behave correctly for subscriptions started on the main thread (e.g., view model RACAbles), since we have a known scheduler there.

As part of this, I had to remove the +subscriptionScheduler cancellation test, since there's no way to reliably test it if we can't prevent it from running blocks as soon as they're scheduled. I don't consider this a big loss, since it's covered by all the other cancellation tests anyways.