Create new documentation for Obj-C operators.
I spent much time until finally understood how exactly do basic operators work and why we need them. This post was intended for my blog, but maybe it can be useful for other newcomers here, in documentation. You are welcome to reject the patch if this information is not appropriate. Thanks!