Remove inclusion of config.h
Other libraries define the HAVE_CONFIG_H condition to include
config.h, such as LibYAML or YAML.framework. When this happens
ReactiveCocoa’s extobjc tries to include its own config.h file which
doesn't
exist. As this doesn’t seem to be used in ReactiveCocoa, we can safely
remove it.