Fix swizzled implementation of `-methodSignatureForSelector:`
The issue is when KVO swizzled (or any additional dynamic subclassing) that takes place _after_ RAC has already swizzled the class, when the original class implements message forwarding.

By using the superclass of the actualClass, (actualClass being the second swizzled class), super dispatch can kick into an infinite loop.

We now start super dispatch with the superclass of the RAC swizzled class to avoid this.