Update error handling of +try: according to Cocoa conventions
Whether or not a method fails should be based on the return value of the
method rather than whether an NSError pointer has been set. If the
return value is nil, it is assumed to have failed.

Thus nil is no longer a possible value on the signal returned by `+try:`.