Use availability to restore old functionality
For consumers of the API that are still targeting OS X 10.9, we still need to offer the non-QOS variant of the QueueScheduler API. Using the new availability macros introduced in Swift 2, we can restore the old functionality that 10.9 consumers may rely on.

Note that this changeset also affects some of the schemes to build the tests using the Debug configuration so that we can use the @testable directive to expose internal function definitions to the unit tests.