[POC] Reconsider associatedProperty(), associatedObject(), rex_valueProperty() interfaces
This change enables to pass the host (receiver) instance to `initial` and `setter` closures. Thus the caller of the functions don't need to use weak-captured host instance to implement the closures.

`NSObjectType` protocol and `rex_valueProperty()` which is basend on its `InstanceType` typealias might seem tricky. We can remove the trick if needed with the way as `UIBarItem.rex_enabled` is implemented (direct use of `associatedProperty()`).