Move ReactiveCocoa files into /Sources.
Without doing this, the Swift Package Manager
assumes that /Carthage is a package, and attempts
to build it.

Ideally there’d be a way to ignore directories in
Package.swift, but it doesn’t appear that it’s
possible at the moment, so this will have to do.

“carthage build --no-skip-current” and normal
Xcode build/test still work as expected. The
workspace has been left in the root directory,
so it’s not harder to work on the project.