[WIP] User Interface Recipes
Please note that this document is *very much* a work in progress. Some things that I've left to do:

- [ ] Remove the dependency on https://github.com/neilpa/Rex
- [ ] Produce an actual example project. (Much of the below code was typed "raw" into my text editor, and likely won't compile.)
- [ ] Get feedback from people who have been doing this much longer
- [ ] Fix the TODOs

## Motivation / Author Notes (To remove for final docs, if deemed inappropriate.)

@liscio - Jan 2016

There really isn't any sort of "canonical document" for creating UI code using ReactiveCocoa, and that's probably due to the fact that it's difficult to find a single approach or "recipe" that applies to all scenarios. In my own work on [Capo](http://capoapp.com) and [FuzzMeasure](http://fuzzmeasure.com) over the past 6 months I have found so many different scenarios and examples that forced me to repeatedly question what I was doing. 

Just when I think that I have an approach figured out, I find another handful of new challenges when I move to another part of the UI to work on. For instance, when I built a lot of iOS UI code I found that it was very easy to build static UIs for `UITableViewCell`s and use `ViewModel`s to ensure that I encapsulate the presentation behavior.

Once I moved to editable controls for the [Neptune](http://capoapp.com/neptune) UI I faced a whole new set of challenges when trying to get values propagated from my ViewModel down into my audio engine, and encapsulating a live-updating visualization (that wasn't needlessly updating the screen while playback wasn't active, also!)

I feel that this collection of "Recipes" for various UI patterns would be extremely helpful to people that are new to ReactiveCocoa, and folks that might be looking for ideas when discovering new challenges that they haven't faced before.