Add type-safe method lifting with weak reference semantics
Introduces a new `MethodType` protocol and concrete `Method` class.  A
`Method` encapsulates a weak reference to an object, and a function that
accepts a reference to the object, some input, and returns the method's
output.

When a `Method` is lifted using a signal or signal producer of input, it
returns a signal or signal producer (respectively) of output from the
results of calling the function with the input.

If the weakly-referenced object is deallocated, the lifted signal or
signal producer will send an interrupted event. When the Objective-C
runtime is present, this is achieved by observing the object's
deallocation at runtime. There is a fallback implementation for native
Swift objects: even though it's impossible to be notified when a
weakly-referenced Swift object is deallocated, we can wait until the
next input value is sent and lazily interrupt the signal if the result
of the method call is `nil`.