Avoid sending events after disposable have been disposed.
The API of `RACPassthroughSubscriber` mentions that no more events will
be sent after the disposable have been disposed. However, this is not
correct if the disposable is disposed on a different thread than the
thread that sends the events, as race conditions are possible inside the
send[Next|Completed|Error] methods, after the `disposable.disposed`
test.